scaffolding for returning exceptions to js + fix emcc -o2 bug
emcc -o2 and -o3 (unlike -o1 -o0, or any clang) had a problem where if I assigned a struct member
a string or u16string constructed using a char* or char16_t* respectfully, and did not use or referenece
the struct member or array for the rest of the function and returned the struct, a \u0011 terminator value
would get appended to the end of the string, obviously causing problems elsewhere. again it's a heisenbug,
so if you print the array or member value within the same function that fixes the problem too. It's only
happened to me so far with arrays (vs pointers) and with 2-length arrays so the case could be limited to that
in either way the workaround for emcc to not give trouble is to just initialize a separate array in-place