another value that needs to be initialized to null, see notes
The selective application of default null values to Tokens and Nodes is discomfiting to say the least
(in the original esprima.js, it happens almost purely as a formality)
here in the C++ it's been done on an as-needed basis and there's a reasoning behind it in that
doing so required the annotation and making explicit during development of different kinds of
null and false values. Also, there are some assignations of NULL explicitly in the originally,
which would not be the default state for the js vars (default is the value undefined) so again,
isNull is this explicit representation of that and it may not be 1:1 to make Nodes Null by default.

that is: usually Nodes are checked against NULL not undefined in original.

However, there are some contexts where Tokens are checked against undefined, and it seems kind of
arbitrary or at least not ideally semantically rich to just set select tokens to null at the beginning
of a function. If they can all safely default to null, that would be better. Test that later.