c++-functional and faster approach to identifying nonascii identifier components
not sure the js regex was even valid (ie. not sure if c++ regexes accepted ranges) but the regex even with
converting \x to \u00 was not working, and it was simply a better return on time to write an approach that would
be plain faster. I can understand for js, where each operation is costly, where a processed-at-native-speed regexp
is a great choice; c++ plain old data structures are fine.